refactor(proto): move FromMillis/ToMillis to separate mod
